Serializer: Factor-out slice writing to SerializerBackend trait impl
This allows implementing different writing backends.
